backend/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
pandas==1.3.3
numpy==1.21.2
requests==2.26.0
yfinance==0.2.12
python-dotenv==0.19.0
fastapi==0.108.0
uvicorn==0.25.0
cachetools==5.3.0
matplotlib==3.7.1
plotly==5.18.0
beautifulsoup4==4.12.2
jb-news==1.0.0 